add example with every test timeout
