+++ examples/fundamentals__timeout/cypress.json
@@ -1,5 +1,6 @@
 {
   "fixturesFolder": false,
   "pluginsFile": false,
-  "supportFile": false
+  "supportFile": false,
+  "testFiles": "**/*spec.js"
 }
